Update patch set 1
Patch Set 1:

The ties into ril seem to be manageable (according to ida), it's one big function which reacts to RIL (rds) messages.

But my first goal was to fix stuttering SCO-audio (when audio is generated by the AP, not in calls, it's fine there), but this approach leads to exact the same brokenness as the closed-source HAL.

I know it's not the kernel-driver because playing audio on the right device works fine and now I know it is not the HAL...

But regardless of this, I will continue working on this change.

Patch-set: 1