fixboot: Fix handling of empty SLOT_LOC
If mmcblk1p25 emstorage is formatted as ext4 or something other than
vfat, booting the stock partition fails because we wrongly try mount
the loopback devices. This happens because SLOT_LOCK is empty
and we still pass the "$SLOT_LOC" != "stock" test.

Let's fix the issue by bailing out early if SLOT_LOC is empty.