Update patch set 1
Patch Set 1:

Sorry ran out of disk space trying to do repo sync and I could not figure out how to add the patch via the web interface. So if possible, please use the the github pull request instead as I also updated the comments a bit.

The pull request is at:
https://github.com/LineageOS/android_device_motorola_omap4-common/pull/2

Patch-set: 1
CC: Tony Lindgren <20663@1ec6b3db-b2c1-4fa4-84a2-4c7efe89ba71>